requirements.txt
pandas
matplotlib
Pillow
bs4